# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Job Board Tech API è uma API para gerenciamento de vagas de emprego.",
        "title": "Job Board Tech API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/auth/": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "User data to register",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login with email and password",
                "parameters": [
                    {
                        "description": "User data to login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login response",
                        "schema": {
                            "$ref": "#/definitions/application.SessionOut"
                        }
                    }
                }
            }
        },
        "/auth/login/oauth": {
            "get": {
                "description": "Redirect to login with oauth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Redirect to login with oauth",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Oauth provider",
                        "name": "oauth_provider",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/login/oauth/callback": {
            "get": {
                "description": "Callback oauth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Callback oauth",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login response",
                        "schema": {
                            "$ref": "#/definitions/application.SessionOut"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "application.SessionOut": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "data_expiracao": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "application.UserRegisterReq": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}